bigquery_tables/requirements.txt
google.cloud==0.34.0
google-cloud-bigquery==3.11.4